Add faction reputation module and tests
